+++ src/WatchCompass.Application/Application/DependencyInjection/ApplicationServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using WatchCompass.Application.UseCases.Recommendations;
+using WatchCompass.Application.UseCases.MovieDetails;
 using WatchCompass.Application.UseCases.SearchMovies;
 
 namespace WatchCompass.Application.DependencyInjection;
@@ -10,6 +11,7 @@ public static class ApplicationServiceCollectionExtensions
     {
         services.AddScoped<SearchMoviesUseCase>();
         services.AddScoped<GetRecommendationsUseCase>();
+        services.AddScoped<GetMovieDetailsUseCase>();
         return services;
     }
 }