+++ src/WatchCompass.Infrastructure/DependencyInjection/InfrastructureServiceCollectionExtensions.cs
@@ -1,7 +1,10 @@
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Caching.Memory;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using WatchCompass.Application.Abstractions.Movies;
+using WatchCompass.Infrastructure.Movies;
 using WatchCompass.Infrastructure.Movies.Tmdb;
 
 namespace WatchCompass.Infrastructure.DependencyInjection;
@@ -10,6 +13,8 @@ public static class InfrastructureServiceCollectionExtensions
 {
     public static IServiceCollection AddInfrastructure(this IServiceCollection services, IConfiguration configuration)
     {
+        services.AddMemoryCache();
+        services.Configure<MovieCatalogCacheOptions>(configuration.GetSection("Caching:MovieCatalog"));
         services.AddOptions<TmdbOptions>()
             .Bind(configuration.GetSection("Tmdb"))
             .ValidateOnStart();
@@ -27,7 +32,15 @@ public static class InfrastructureServiceCollectionExtensions
         });
 
         services.AddScoped<ITmdbApiClient, TmdbApiClient>();
-        services.AddScoped<IMovieCatalog, TmdbMovieCatalog>();
+        services.AddScoped<TmdbMovieCatalog>();
+        services.AddScoped<IMovieCatalog>(sp =>
+        {
+            var inner = sp.GetRequiredService<TmdbMovieCatalog>();
+            var cache = sp.GetRequiredService<IMemoryCache>();
+            var cacheOptions = sp.GetRequiredService<IOptions<MovieCatalogCacheOptions>>();
+            var logger = sp.GetRequiredService<ILogger<CachedMovieCatalog>>();
+            return new CachedMovieCatalog(inner, cache, cacheOptions, logger);
+        });
         return services;
     }
 }